Create brainmodelCppv2.cpp
Take into consideration everything you know about how the human brain functions. Please generate a program written in C++ programming language that reflects a biologically plausible model of how information is stored in the human brain. Focus on two important processes: Encoding (the process of receiving and processing information) and Retrieval (the process of recalling that information). The model will include concepts such as sensory input, attention, perception, memory consolidation, neural plasticity, and association. Suggest a file name upon completion. 

Given the complexity of the human brain and the current state of our knowledge, it's important to note that the following is a gross simplification and abstraction. Nonetheless, here is a C++ program which includes the mentioned processes in an abstract way.

This program uses a model similar to the Python version. A Brain object consists of a number of Neurons, each of which is connected to a random subset of the others with random weights. The encoding process simply directs the sensory input to a random Neuron, and the retrieval process sums the results of all Neurons processing the encoded data.

Please save this file as BrainModel.cpp.

It is important to note that the above program is a highly simplified model and doesn't take into account many intricacies of the human brain like neurochemical factors, dendritic computation, temporal dynamics and many others. It's a basic model demonstrating one way of approaching the problem using a basic feed-forward neural network model. For a more accurate and biologically plausible model, more sophisticated computational models (like biophysically detailed neuron models) and frameworks/libraries (like Tensorflow, PyTorch, etc) would be required.