Create BrainModel.cbl
Take into consideration everything you know about how the human brain functions. Please generate a program written in COBOL programming language that reflects a biologically plausible model of how information is stored in the human brain. Focus on two important processes: Encoding (the process of receiving and processing information) and Retrieval (the process of recalling that information). The model will include concepts such as sensory input, attention, perception, memory consolidation, neural plasticity, and association. Suggest a file name upon completion. 


A program in COBOL that reflects the biological processes of the human brain can be somewhat tricky due to COBOL's original design for business and finance calculations. However, we can simulate the encoding and retrieval processes of information storage. Here's a simple COBOL program that encodes input into memory and retrieves it.

In the above code, an array of 10 elements named SENSORY-INPUT is declared to receive the sensory input data, and another array ENCODED-DATA is used to store the encoded data (the actual information). In this simple model, encoding is done just by copying the sensory input data to the ENCODED-DATA array, and retrieval is simply displaying the encoded data.

This doesn't capture the complexity of the human brain's processes, but it can serve as a basis for a more complex model. Save this file as BrainModel.cbl.