Create BrainModel.nim
Take into consideration everything you know about how the human brain functions. Please generate a program written in Nim programming language that reflects a biologically plausible model of how information is stored in the human brain. Focus on two important processes: Encoding (the process of receiving and processing information) and Retrieval (the process of recalling that information). The model will include concepts such as sensory input, attention, perception, memory consolidation, neural plasticity, and association. Suggest a file name upon completion. 

In the Nim programming language, here's a simplified and abstract model of some processes related to information storage in the human brain:

This code first takes in sensory input data, emulates the encoding process by assigning the sensory input to a set of neurons, and then retrieves the encoded information. As a simple model of "neural plasticity", the encoded data for each neuron is also influenced by the encoded data of the neighbouring neurons.

Please save this file as BrainModel.nim.

Keep in mind that this model is an oversimplification and doesn't accurately represent the complexity and functionality of the human brain, but it should give a basic idea of how information might be encoded and retrieved.