Create BrainmodelErlangv2.erl
Take into consideration everything you know about how the human brain functions. Please generate a program written in Erlang Programming Language that reflects a biologically plausible model of how information is stored in the human brain. Focus on two important processes: Encoding (the process of receiving and processing information) and Retrieval (the process of recalling that information). The model will include concepts such as sensory input, attention, perception, memory consolidation, neural plasticity, and association. Suggest a file name upon completion. 

Erlang is a concurrent programming language suitable for distributed systems, but is not typically used for modeling biological systems such as the human brain. Nonetheless, I can provide a basic and highly abstracted simulation of brain activity that involves encoding and retrieving information. I will use Erlang's process messaging capabilities to model the communication between neurons.

Please keep in mind that the actual mechanisms of encoding and retrieving information in the human brain are much more complex, involving changes in synaptic strengths, intricate neural networks, and processes that we still don't fully understand.

Here is a simplistic model:

In this code, the start/0 function spawns two neurons and connects them. It then sends a store message to Neuron1 to simulate encoding a memory (in this case, the memory is simply a symbol, perception). After a delay, it sends a retrieve message to Neuron1 to retrieve the memory. Neuron1 sends the memory back to the calling process, which then prints it out.

Each neuron/2 function represents a neuron. The neuron can store information (the store clause in the receive statement) and can retrieve information (the retrieve clause in the receive statement). The neuron is also capable of connecting with another neuron (the connect clause in the receive statement), simulating a basic form of neural connectivity.

You can save this code to a file called BrainModel.erl.

This Erlang code abstracts the complex neural processes of the human brain into simple message passing between Erlang processes. It's an oversimplification and doesn't model the exact mechanisms of memory encoding and retrieval, but it gives a small glimpse of how concurrent processes can interact to store and retrieve information.