Create tfvariational_autoencoder.py
Here's the basic idea, using the typical notation: we have some data, X, and we're going to pretend that a very simplified model, p, generated it stochastically. This model has a latent variable, z (a vector), that can take on different values, with probabilities p(z). For a given value of z it generates, the model can then generate data points x with probabilities p(x|z). So if you know the correct distribution of p(z) and p(x|z), you can generate data that's similar to/includes X. Typically, people choose the dimension of the latent space to be smaller than dimension of the data, because this 1) compresses the data, and 2) hopefully forces the latent space to learn a representation of the data that captures core features of it. However, this is abstract. The actual goal is to approximate this model with an "encoder", q(z|x), to translate data to that underlying latent space we're assuming, and a "decoder" p(x|z), that lets you produce data from latent values. That's the general approach and could be done in several ways. However, a common way (and what we'll do here) is to approximate the model using a specific NN structure with two (or, depending on how you want to count it, three) main parts: 1) the encoder takes the input x and transforms it into a vector of [μ,σ], 2) These are combined into the latent variable z=μ+σϵ, where ϵ=U(0,1), and 3) the decoder takes this z and produces the output y. The loss term (see the paper for why this makes sense) is a sum of the "reconstruction error", which is just the MSE between x and ^x, and the Kullback-Leibler divergence between the prior p(z) (which is usually taken to be N(0,1) with the dimensions of z), and the encoder distribution q(z|x). Using this information generate a variational autoencoder using tensorflow and suggest a file name upon completion.

Below is a simple Python script for implementing a Variational Autoencoder (VAE) using TensorFlow and Keras. The script creates a VAE model, compiles it with a loss function that includes both reconstruction error and KL divergence, and sets up a dummy training loop.

This example creates a very basic VAE with a single hidden layer in the encoder and decoder, and it uses a binary cross entropy loss for the reconstruction error. In practice, you would adjust the architecture and loss function to suit your specific problem. You'd also replace the dummy training loop with actual training data.

You can save this script as variational_autoencoder.py. Please note, this is a basic version of VAE and in real-world scenarios, you would have to modify it to fit your needs.